+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/bean/DozerBeanTransformer.java
@@ -10,8 +10,16 @@ import java.util.Objects;
 /**
  * 使用 Dozer 实现的 Bean 映射器。
  *
+ * <p>
+ * 由于 <code>Dozer</code> 框架已经停止维护，故此实现过期，请勿再使用。
+ * <p>
+ * 推荐使用 <code>MapStruct</code> 框架替代 <code>Dozer</code> 框架。后者无论安全性还是效率都高于前者。<br>
+ * 相关的实现: {@link MapStructBeanTransformer}。
+ *
  * @author DwArFeng
+ * @see MapStructBeanTransformer
  * @since 0.0.1-alpha
+ * @deprecated 由于 <code>Dozer</code> 框架已经停止维护，故此实现过期，请勿再使用。
  */
 public class DozerBeanTransformer<U extends Bean, V extends Bean> implements BeanTransformer<U, V> {
 
@@ -20,9 +28,8 @@ public class DozerBeanTransformer<U extends Bean, V extends Bean> implements Bea
     private Mapper mapper;
 
     public DozerBeanTransformer(
-            @NonNull Class<U> classU,
-            @NonNull Class<V> classV,
-            @NonNull Mapper mapper) {
+            @NonNull Class<U> classU, @NonNull Class<V> classV, @NonNull Mapper mapper
+    ) {
         this.classU = classU;
         this.classV = classV;
         this.mapper = mapper;