+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/bean/IdentityBeanTransformer.java
@@ -4,9 +4,10 @@ import com.dwarfeng.subgrade.stack.bean.Bean;
 import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
 
 /**
- * 本征 BeanMapper。
+ * 本征 BeanTransformer。
  *
- * <p>将对象本身映射为自己的 BeanMapper</p>
+ * <p>
+ * 将对象本身映射为自己的 BeanTransformer。
  *
  * @author DwArFeng
  * @since 0.0.1-alpha